requirements.txt
Flask==3.0.0
Flask_SocketIO==5.3.6
python-dotenv==1.0.0
